Add new protocol of the communication of the clint and server

+++ HitProject/ConnectStatesCell.h
@@ -22,7 +22,7 @@
 @property (nonatomic) NSString *robotVoice;//50
 
 @property (nonatomic) NSString *robotTemPower;//27.2v
-@property (nonatomic) int      times;//累加次数
+@property (nonatomic) NSInteger times;//累加次数
 @property (nonatomic) NSMutableArray *multPowerArray;//电量数组
 
 @end
